hmm i think this should do it

+++ examples/distance.c
@@ -25,6 +25,7 @@
 #include <h3/h3api.h>
 #include <math.h>
 #include <stdio.h>
+#include <inttypes.h>
 
 /**
  * @brief haversineDistance finds the
@@ -65,10 +66,10 @@ int main(int argc, char *argv[]) {
     printf(
         "origin: (%lf, %lf)\n"
         "destination: (%lf, %lf)\n"
-        "grid distance: %ld\n"
+        "grid distance: %" PRId64 "\n"
         "distance in km: %lfkm\n",
         radsToDegs(geoHQ1.lat), radsToDegs(geoHQ1.lng), radsToDegs(geoHQ2.lat),
-        radsToDegs(geoHQ2.lng), (long)distance,
+        radsToDegs(geoHQ2.lng), distance,
         haversineDistance(geoHQ1.lat, geoHQ1.lng, geoHQ2.lat, geoHQ2.lng));
     // Output:
     // origin: (37.775236, -122.419755)
